+++ src/test/java/lotto/domain/dao/WinningLottoDaoTest.java
@@ -20,7 +20,7 @@ class WinningLottoDaoTest {
         LottoTicket lottoTicket = new LottoTicket(Arrays.asList(1, 2, 3, 4, 5, 6));
         int bonusBall = 7;
         WinningLotto winningLotto = new WinningLotto(lottoTicket, bonusBall);
-        winningLottoDao.addWinningLotto(winningLotto);
+        winningLottoDao.addWinningLotto(1, winningLotto);
     }
 
     @Test
